完成了App_IO文件，包含按键处理的流程，不同按键输入输出不同的按键响应

+++ main/Dri/Dri_NVS.h
@@ -11,7 +11,7 @@ void Dri_NVS_Init(void);
 
 esp_err_t Dri_NVS_WriteStr(uint8_t key[], uint8_t value[]);
 
-esp_err_t Dri_NVS_ReadStr(uint8_t key[], uint8_t value[],size_t * len);
+esp_err_t Dri_NVS_ReadStr(uint8_t key[], uint8_t value[],uint8_t *len);
 
 esp_err_t Dri_NVS_DeleKey(uint8_t key[]);
